+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/blueFront.java
@@ -1,6 +1,8 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.autos;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+import org.firstinspires.ftc.teamcode.drive.RobotEncoded;
 
 @Autonomous
 public class blueFront extends LinearOpMode {
